Moved _eval_input to utils.
